refactor: get API에 따른 로직 수정
